refactor: extract bp_crons to routes/crons.py
Phase 5.4 of the incremental modularisation. Moves the 13 cron CRUD +
health + run-log endpoints out of the 31k-line dashboard.py into a
dedicated routes/crons.py blueprint module.

Routes moved:
- GET  /api/crons
- POST /api/cron/fix
- POST /api/cron/run
- POST /api/cron/toggle
- POST /api/cron/delete
- POST /api/cron/update
- POST /api/cron/create
- GET  /api/cron/<job_id>/runs
- POST /api/cron/<job_id>/kill
- GET  /api/cron-run-log
- GET  /api/cron/health-summary
- POST /api/cron/kill-all
- GET  /api/cron-health

Helpers (_enrich_cron_runs, _cron_runs_from_transcripts, _gw_invoke,
_get_crons, _compute_transcript_analytics, _score_cron_match, _ext_emit,
_get_sessions_dir, app, _budget_paused) stay in dashboard.py and are
reached via late `import dashboard as _d` inside each handler — same
pattern as routes/sessions.py, routes/brain.py, routes/usage.py.

Pure mechanical move, zero behaviour change.

dashboard.py: 31729 -> 31260 lines (-469)
routes/crons.py: 529 lines (new)
